TASK-54243 : search news by tags (#453)

+++ services/src/main/java/org/exoplatform/news/listener/NewsMetadataListener.java
@@ -1,5 +1,6 @@
 package org.exoplatform.news.listener;
 
+import org.apache.commons.collections.MapUtils;
 import org.apache.commons.lang3.StringUtils;
 import org.exoplatform.commons.search.index.IndexingService;
 import org.exoplatform.news.model.News;
@@ -18,8 +19,15 @@ import org.exoplatform.social.core.processor.MetadataActivityProcessor;
 import org.exoplatform.social.metadata.favorite.FavoriteService;
 import org.exoplatform.social.metadata.favorite.model.Favorite;
 import org.exoplatform.social.metadata.model.MetadataItem;
+import org.exoplatform.social.metadata.tag.TagService;
+import org.exoplatform.social.metadata.tag.model.TagName;
+import org.exoplatform.social.metadata.tag.model.TagObject;
+
+import java.util.Set;
 
 public class NewsMetadataListener extends Listener<Long, MetadataItem> {
+
+  public static final String NEWS_METADATA_OBJECT_TYPE = "news";
 
   private final IndexingService indexingService;
 
@@ -31,20 +39,30 @@ public class NewsMetadataListener extends Listener<Long, MetadataItem> {
 
   private final ActivityManager activityManager;
 
+  private TagService tagService;
+
+
   private static final String   METADATA_CREATED = "social.metadataItem.created";
 
   private static final String   METADATA_DELETED = "social.metadataItem.deleted";
+
+  private static final String   METADATA_TAG = "tags";
+
+  private static final String   METADATA_FAVORITE = "favorite";
+
 
   public NewsMetadataListener(IndexingService indexingService,
                               NewsService newsService,
                               FavoriteService favoriteService,
                               IdentityManager identityManager,
-                              ActivityManager activityManager) {
+                              ActivityManager activityManager,
+                              TagService tagService) {
     this.indexingService = indexingService;
     this.newsService = newsService;
     this.favoriteService = favoriteService;
     this.identityManager = identityManager;
     this.activityManager = activityManager;
+    this.tagService = tagService;
   }
 
   @Override
@@ -69,12 +87,53 @@ public class NewsMetadataListener extends Listener<Long, MetadataItem> {
                                          "",
                                          Long.parseLong(userIdentity.getId()));
         if (event.getEventName().equals(METADATA_CREATED)) {
-          favoriteService.createFavorite(favorite);
-        } else if (event.getEventName().equals(METADATA_DELETED)) {
+          if (!metadataItem.getObjectType().equals(NEWS_METADATA_OBJECT_TYPE) && metadataItem.getMetadataTypeName().equals(METADATA_TAG)) {
+            updateActivityTags(activity, news);
+          }
+          else {
+            favoriteService.createFavorite(favorite);
+          }
+        } else if (event.getEventName().equals(METADATA_DELETED) && metadataItem.getMetadataTypeName().equals(METADATA_FAVORITE)) {
           favoriteService.deleteFavorite(favorite);
         }
         indexingService.reindex(NewsIndexingServiceConnector.TYPE, news.getId());
       }
     }
   }
+  private void updateActivityTags(ExoSocialActivity activity, News news) {
+    String objectType = NEWS_METADATA_OBJECT_TYPE;
+
+    long creatorId = getPosterId(activity);
+
+    org.exoplatform.social.core.identity.model.Identity audienceIdentity = activityManager.getActivityStreamOwnerIdentity(activity.getId());
+    long audienceId = Long.parseLong(audienceIdentity.getId());
+    String content = getActivityBody(activity);
+
+    Set<TagName> tagNames = tagService.detectTagNames(content);
+    tagService.saveTags(new TagObject(objectType,
+                    news.getId(),
+                    activity.getParentId()),
+            tagNames,
+            audienceId,
+            creatorId);
+  }
+  private long getPosterId(ExoSocialActivity activity) {
+    String userId = activity.getUserId();
+    if (StringUtils.isBlank(userId)) {
+      userId = activity.getPosterId();
+    }
+    return StringUtils.isBlank(userId) ? 0 : Long.parseLong(userId);
+  }
+
+  private String getActivityBody(ExoSocialActivity activity) {
+    String body = MapUtils.getString(activity.getTemplateParams(), "comment");
+    if (StringUtils.isNotBlank(body)) {
+      return body;
+    } else if (StringUtils.isNotBlank(activity.getTitle())) {
+      return activity.getTitle();
+    } else {
+      return activity.getBody();
+    }
+  }
+
 }
